DR3256AC-914 Ruff and mypi fix.

+++ ic_haus_magnetic_encoder_calibration/encoder.py
@@ -148,8 +148,8 @@ class CalibrationResult:
     nonius_adjustments: Optional[mu_3sl.AnalogTrackAdjustments] = None
     spo_base: int = 0
     spo_n: list[int] = field(default_factory=list)
-    nonius_in_range_max: int = 0
-    nonius_in_range_min: int = 0
+    nonius_in_range_max: float = 0.0
+    nonius_in_range_min: float = 0.0
 
 
 # ---------------------------------------------------------------------------
